backend/storage-access-function/requirements.txt
functions-framework==3.*
google-cloud-storage==3.2.0
Flask==3.0.*
firebase-admin==6.2.0
python-dotenv==1.1.1
